Extract DistrictMap into separate component
